Code Refactored for guest service
