+++ guest-service/src/main/java/com/epam/guest/GuestApplication.java
@@ -1,4 +1,4 @@
-package com.epam.guestservice;
+package com.epam.guest;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -9,10 +9,10 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2;
 @SpringBootApplication
 @EnableEurekaClient
 @EnableSwagger2
-public class GuestserviceApplication {
+public class GuestApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(GuestserviceApplication.class, args);
+		SpringApplication.run(GuestApplication.class, args);
 	}
 
 }
