+++ guest-service/src/main/java/com/epam/guest/configuration/SwaggerConfiguration.java
@@ -1,4 +1,4 @@
-package com.epam.guestservice.configuration;
+package com.epam.guest.configuration;
 
 import java.util.ArrayList;
