+++ guest-service/src/main/java/com/epam/guest/entity/Profile.java
@@ -1,4 +1,4 @@
-package com.epam.guestservice.entity;
+package com.epam.guest.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
