+++ guest-service/src/main/java/com/epam/guest/model/CreditCard.java
@@ -1,4 +1,4 @@
-package com.epam.guestservice.model;
+package com.epam.guest.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
