+++ guest-service/src/main/java/com/epam/guest/model/User.java
@@ -1,4 +1,4 @@
-package com.epam.guestservice.model;
+package com.epam.guest.model;
 
 import java.util.List;
