+++ guest-service/src/main/java/com/epam/guest/repository/GuestRepository.java
@@ -1,0 +1,9 @@
+package com.epam.guest.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.epam.guest.entity.User;
+
+public interface GuestRepository extends JpaRepository<User, Integer> {
+
+}
